refactor: move routes to http/handlers
