+++ crates/saldo-ws/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod http;
-pub mod routes;
 pub use sblex_telemetry as telemetry;
